Switching collision masks & groups to use long instead of int.

+++ atlas/src/main/java/us/ihmc/atlas/parameters/AtlasSimulationCollisionModel.java
@@ -60,8 +60,8 @@ public class AtlasSimulationCollisionModel implements RobotCollisionModel
    public List<Collidable> getRobotCollidables(MultiBodySystemBasics multiBodySystem)
    {
       List<Collidable> collidables = new ArrayList<>();
-      int collisionMask = helper.getCollisionMask(robotCollisionMask);
-      int collisionGroup = helper.createCollisionGroup(otherCollisionMasks);
+      long collisionMask = helper.getCollisionMask(robotCollisionMask);
+      long collisionGroup = helper.createCollisionGroup(otherCollisionMasks);
 
       { // Torso
          RigidBodyBasics torso = RobotCollisionModel.findRigidBody(jointMap.getChestName(), multiBodySystem);
